+++ src/anyio/_backends/_trio.py
@@ -132,16 +132,15 @@ class TaskGroup(abc.TaskGroup):
         finally:
             self._active = False
 
-    def spawn(self, func: Callable, *args, name=None) -> DeprecatedAwaitable:
+    def start_soon(self, func: Callable, *args, name=None) -> None:
         if not self._active:
-            raise RuntimeError('This task group is not active; no new tasks can be spawned.')
+            raise RuntimeError('This task group is not active; no new tasks can be started.')
 
         self._nursery.start_soon(func, *args, name=name)
-        return DeprecatedAwaitable(self.spawn)
 
     async def start(self, func: Callable[..., Coroutine], *args, name=None):
         if not self._active:
-            raise RuntimeError('This task group is not active; no new tasks can be spawned.')
+            raise RuntimeError('This task group is not active; no new tasks can be started.')
 
         return await self._nursery.start(func, *args, name=name)
 
@@ -164,6 +163,9 @@ run_sync_from_thread = trio.from_thread.run_sync
 
 
 class BlockingPortal(abc.BlockingPortal):
+    def __new__(cls):
+        return object.__new__(cls)
+
     def __init__(self):
         super().__init__()
         self._token = trio.lowlevel.current_trio_token()
@@ -171,7 +173,7 @@ class BlockingPortal(abc.BlockingPortal):
     def _spawn_task_from_thread(self, func: Callable, args: tuple, kwargs: Dict[str, Any],
                                 name, future: Future) -> None:
         return trio.from_thread.run_sync(
-            partial(self._task_group.spawn, name=name), self._call_func, func, args, kwargs,
+            partial(self._task_group.start_soon, name=name), self._call_func, func, args, kwargs,
             future, trio_token=self._token)