+++ src/anyio/_core/_compat.py
@@ -84,6 +84,9 @@ class DeprecatedAwaitable:
         if False:
             yield
 
+    def __reduce__(self):
+        return type(None), ()
+
     def _unwrap(self):
         return None
 
@@ -102,6 +105,9 @@ class DeprecatedAwaitableFloat(float):
 
         return float(self)
 
+    def __reduce__(self):
+        return float, (float(self),)
+
     def _unwrap(self) -> float:
         return float(self)
 
@@ -117,6 +123,9 @@ class DeprecatedAwaitableList(List[T]):
             yield
 
         return self
+
+    def __reduce__(self):
+        return list, (list(self),)
 
     def _unwrap(self) -> List[T]:
         return list(self)