+++ src/anyio/_core/_synchronization.py
@@ -4,7 +4,7 @@ from types import TracebackType
 from typing import Deque, Optional, Tuple, Type
 from warnings import warn
 
-from ..lowlevel import checkpoint
+from ..lowlevel import cancel_shielded_checkpoint, checkpoint, checkpoint_if_cancelled
 from ._compat import DeprecatedAwaitable
 from ._eventloop import get_asynclib
 from ._exceptions import BusyResourceError, WouldBlock
@@ -114,7 +114,7 @@ class Lock:
 
     async def acquire(self) -> None:
         """Acquire the lock."""
-        await checkpoint()
+        await checkpoint_if_cancelled()
         try:
             self.acquire_nowait()
         except WouldBlock:
@@ -131,6 +131,8 @@ class Lock:
                 raise
 
             assert self._owner_task == task
+        else:
+            await cancel_shielded_checkpoint()
 
     def acquire_nowait(self) -> None:
         """
@@ -289,6 +291,7 @@ class Semaphore:
 
     async def acquire(self) -> None:
         """Decrement the semaphore value, blocking if necessary."""
+        await checkpoint_if_cancelled()
         try:
             self.acquire_nowait()
         except WouldBlock:
@@ -301,8 +304,8 @@ class Semaphore:
                     self._waiters.remove(event)
 
                 raise
-
-            self.acquire_nowait()
+        else:
+            await cancel_shielded_checkpoint()
 
     def acquire_nowait(self) -> None:
         """
@@ -321,9 +324,10 @@ class Semaphore:
         if self._max_value is not None and self._value == self._max_value:
             raise ValueError('semaphore released too many times')
 
-        self._value += 1
         if self._waiters:
             self._waiters.popleft().set()
+        else:
+            self._value += 1
 
         return DeprecatedAwaitable(self.release)