+++ tests/test_compat.py
@@ -1,3 +1,4 @@
+import pickle
 import signal
 import sys
 import threading
@@ -6,10 +7,13 @@ import pytest
 
 from anyio import (
     CancelScope, CapacityLimiter, Condition, Event, Lock, Semaphore, TaskInfo,
-    create_memory_object_stream, create_task_group, current_default_worker_thread_limiter,
-    current_effective_deadline, current_time, fail_after, get_current_task, get_running_tasks,
-    maybe_async, maybe_async_cm, move_on_after, open_signal_receiver, run_async_from_thread,
-    run_sync_from_thread, run_sync_in_worker_thread, sleep, to_thread)
+    create_blocking_portal, create_memory_object_stream, create_task_group,
+    current_default_worker_thread_limiter, current_effective_deadline, current_time, fail_after,
+    get_current_task, get_running_tasks, maybe_async, maybe_async_cm, move_on_after,
+    open_signal_receiver, run_async_from_thread, run_sync_from_thread, run_sync_in_worker_thread,
+    to_thread)
+from anyio._core._compat import (
+    DeprecatedAwaitable, DeprecatedAwaitableFloat, DeprecatedAwaitableList)
 
 pytestmark = pytest.mark.anyio
 
@@ -125,11 +129,6 @@ class TestDeprecations:
         with pytest.deprecated_call():
             await semaphore.release()
 
-    async def test_taskgroup_spawn(self):
-        async with create_task_group() as tg:
-            with pytest.deprecated_call():
-                await tg.spawn(sleep, 0)
-
     async def test_move_on_after(self):
         with pytest.deprecated_call():
             async with move_on_after(0):
@@ -166,3 +165,27 @@ class TestDeprecations:
         with pytest.deprecated_call():
             default_limiter = to_thread.current_default_thread_limiter()
             assert current_default_worker_thread_limiter() is default_limiter
+
+    async def test_create_blocking_portal(self):
+        with pytest.deprecated_call():
+            async with create_blocking_portal():
+                pass
+
+
+class TestPickle:
+    def test_deprecated_awaitable_none(self):
+        obj = DeprecatedAwaitable(threading.get_ident)
+        result = pickle.loads(pickle.dumps(obj))
+        assert result is None
+
+    def test_deprecated_awaitable_float(self):
+        obj = DeprecatedAwaitableFloat(2.3, threading.get_ident)
+        result = pickle.loads(pickle.dumps(obj))
+        assert type(result) is float
+        assert result == 2.3
+
+    def test_deprecated_awaitable_list(self):
+        obj = DeprecatedAwaitableList([1, 'a'], func=threading.get_ident)
+        result = pickle.loads(pickle.dumps(obj))
+        assert type(result) is list
+        assert result == [1, 'a']