+++ tests/test_from_thread.py
@@ -5,8 +5,9 @@ from contextlib import suppress
 import pytest
 
 from anyio import (
-    Event, create_blocking_portal, from_thread, get_cancelled_exc_class, get_current_task, run,
-    sleep, start_blocking_portal, to_thread, wait_all_tasks_blocked)
+    Event, from_thread, get_cancelled_exc_class, get_current_task, run, sleep, to_thread,
+    wait_all_tasks_blocked)
+from anyio.from_thread import BlockingPortal, start_blocking_portal
 
 pytestmark = pytest.mark.anyio
 
@@ -128,7 +129,7 @@ class TestBlockingPortal:
             thread_ids.append(portal.call(async_get_thread_id))
 
         thread_ids = []
-        async with create_blocking_portal() as portal:
+        async with BlockingPortal() as portal:
             thread = threading.Thread(target=external_thread)
             thread.start()
             await to_thread.run_sync(thread.join)
@@ -148,7 +149,7 @@ class TestBlockingPortal:
 
         results = []
         with suppress(Exception):
-            async with create_blocking_portal() as portal:
+            async with BlockingPortal() as portal:
                 thread1 = threading.Thread(target=external_thread)
                 thread1.start()
                 thread2 = threading.Thread(target=external_thread)
@@ -175,7 +176,7 @@ class TestBlockingPortal:
                 results.append(None)
 
         results = []
-        async with create_blocking_portal() as portal:
+        async with BlockingPortal() as portal:
             thread1 = threading.Thread(target=external_thread)
             thread1.start()
             thread2 = threading.Thread(target=external_thread)
@@ -189,7 +190,7 @@ class TestBlockingPortal:
         assert results == [None, None]
 
     async def test_call_portal_from_event_loop_thread(self):
-        async with create_blocking_portal() as portal:
+        async with BlockingPortal() as portal:
             exc = pytest.raises(RuntimeError, portal.call, threading.get_ident)
             exc.match('This method cannot be called from the event loop thread')
 
@@ -217,7 +218,7 @@ class TestBlockingPortal:
         pytest.raises(RuntimeError, portal.call, threading.get_ident).\
             match('This portal is not running')
 
-    def test_spawn_task(self, anyio_backend_name, anyio_backend_options):
+    def test_start_task_soon(self, anyio_backend_name, anyio_backend_options):
         async def event_waiter():
             await event1.wait()
             event2.set()
@@ -226,23 +227,23 @@ class TestBlockingPortal:
         with start_blocking_portal(anyio_backend_name, anyio_backend_options) as portal:
             event1 = portal.call(Event)
             event2 = portal.call(Event)
-            future = portal.spawn_task(event_waiter)
+            future = portal.start_task_soon(event_waiter)
             portal.call(event1.set)
             portal.call(event2.wait)
             assert future.result() == 'test'
 
-    def test_spawn_task_cancel_later(self, anyio_backend_name, anyio_backend_options):
+    def test_start_task_soon_cancel_later(self, anyio_backend_name, anyio_backend_options):
         async def noop():
             await sleep(2)
 
         with start_blocking_portal(anyio_backend_name, anyio_backend_options) as portal:
-            future = portal.spawn_task(noop)
+            future = portal.start_task_soon(noop)
             portal.call(wait_all_tasks_blocked)
             future.cancel()
 
         assert future.cancelled()
 
-    def test_spawn_task_cancel_immediately(self, anyio_backend_name, anyio_backend_options):
+    def test_start_task_soon_cancel_immediately(self, anyio_backend_name, anyio_backend_options):
         async def event_waiter():
             nonlocal cancelled
             try:
@@ -252,19 +253,19 @@ class TestBlockingPortal:
 
         cancelled = False
         with start_blocking_portal(anyio_backend_name, anyio_backend_options) as portal:
-            future = portal.spawn_task(event_waiter)
+            future = portal.start_task_soon(event_waiter)
             future.cancel()
 
         assert cancelled
 
-    def test_spawn_task_with_name(self, anyio_backend_name, anyio_backend_options):
+    def test_start_task_soon_with_name(self, anyio_backend_name, anyio_backend_options):
         async def taskfunc():
             nonlocal task_name
             task_name = get_current_task().name
 
         task_name = None
         with start_blocking_portal(anyio_backend_name, anyio_backend_options) as portal:
-            portal.spawn_task(taskfunc, name='testname')
+            portal.start_task_soon(taskfunc, name='testname')
 
         assert task_name == 'testname'